% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_perm_results}
\alias{aggregate_perm_results}
\title{Aggregate permutation results into proportions}
\usage{
aggregate_perm_results(results, reps)
}
\arguments{
\item{results}{List of permutation comparison results.}

\item{reps}{Number of permutations.}
}
\value{
A list with proportion matrices \code{lower}, \code{larger},
  \code{abs}.
}
\description{
Aggregate permutation results into proportions
}
\keyword{internal}
