% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_vector.R
\name{array_to_vector}
\alias{array_to_vector}
\title{Internal auxiliary function to transform an array into a vector}
\usage{
array_to_vector(ar, mode., diag.)
}
\arguments{
\item{ar}{array; The \code{array} to be transformed.}

\item{mode.}{character; Is the \code{array} \code{mode. = 'directed'} or \code{mode. = 'undirected'}?}

\item{diag.}{logical; Should the diagonal of each perceiver slice of the array be used?}
}
\value{
Returns a vector
}
\description{
Concatenates the perceiver slices of a 3D array into a single vector, in
the order the rest of the package expects.
}
\keyword{internal}
