% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_internal_formula}
\alias{build_internal_formula}
\title{Build internal formula with additional random intercepts}
\usage{
build_internal_formula(
  formula,
  rin = FALSE,
  ris = FALSE,
  rir = FALSE,
  rip = FALSE
)
}
\arguments{
\item{formula}{A formula object.}

\item{rin}{Logical; random intercept for networks.}

\item{ris}{Logical; random intercept for senders.}

\item{rir}{Logical; random intercept for receivers.}

\item{rip}{Logical; random intercept for perceivers (CSS only).}
}
\value{
An updated formula.
}
\description{
Appends random-intercept terms for sender, receiver, perceiver, network,
and custom grouping to a user-supplied formula.
}
\keyword{internal}
