% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.QAP}
\alias{fitted.QAP}
\alias{residuals.QAP}
\title{Fitted values and residuals from a QAP fit}
\usage{
\method{fitted}{QAP}(object, ...)

\method{residuals}{QAP}(object, ...)
}
\arguments{
\item{object}{A QAP fit.}

\item{...}{Passed to the underlying method.}
}
\value{
A numeric vector.
}
\description{
Taken from the baseline model, i.e. the fit to the unpermuted data.
}
