% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance_QAP}
\alias{glance_QAP}
\title{One-row summary of a QAP fit}
\usage{
glance_QAP(x, ...)
}
\arguments{
\item{x}{A QAP fit.}

\item{...}{Ignored.}
}
\value{
A one-row data frame.
}
\description{
One-row summary of a QAP fit
}
\keyword{internal}
