% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_tab.R
\name{glm_tab}
\alias{glm_tab}
\title{Internal auxiliary function to print QAP() coefficient tables}
\usage{
glm_tab(x, comp)
}
\arguments{
\item{x}{An object of class \code{QAPCSS}.}

\item{comp}{Character or NULL; name of comparison to print.}
}
\value{
Prints the results table.
}
\description{
Internal auxiliary function to print QAP() coefficient tables
}
\keyword{internal}
