% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{gmm_off}
\alias{gmm_off}
\title{Offset from a gmm() data list, defaulting to 0}
\usage{
gmm_off(data)
}
\arguments{
\item{data}{The list passed to \code{gmm::gmm()}.}
}
\value{
Numeric vector, or the scalar 0.
}
\description{
Offset from a gmm() data list, defaulting to 0
}
\keyword{internal}
