% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_utils.R
\name{gpu_available}
\alias{gpu_available}
\title{Check GPU availability}
\usage{
gpu_available()
}
\value{
Logical; \code{TRUE} if both \pkg{torch} and CUDA are available.
}
\description{
Check GPU availability
}
\examples{
gpu_available()
}
