% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{logit_moments}
\alias{logit_moments}
\title{Auxiliary functions for estimation gmm}
\usage{
logit_moments(theta, data)
}
\arguments{
\item{theta}{numeric; Vetor of parameter values, see \code{?gmm::gmm}.}

\item{data}{matrix; Data matrix for all X, see \code{?gmm::gmm}.}

\item{gmmo}{gmm; an object of class gmm.}
}
\value{
Returns moment the moment conditions
}
\description{
Auxiliary functions for estimation gmm
}
