% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{logit_resid}
\alias{logit_resid}
\title{Getting residuals for gmm logit}
\usage{
logit_resid(gmmo)
}
\arguments{
\item{gmmo}{gmm() output}
}
\value{
residuals of the gmm()
}
\description{
Getting residuals for gmm logit
}
