% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{negbin_resid}
\alias{negbin_resid}
\title{Residuals from a GMM negative binomial fit}
\usage{
negbin_resid(gmmo)
}
\arguments{
\item{gmmo}{gmm() output from negbin_moments.}
}
\value{
Numeric vector of residuals.
}
\description{
Residuals from a GMM negative binomial fit
}
\keyword{internal}
