% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{poisson_moments}
\alias{poisson_moments}
\title{Auxiliary functions for GMM estimation - Poisson}
\usage{
poisson_moments(theta, x)
}
\arguments{
\item{theta}{numeric; Vetor of parameter values, see \code{?gmm::gmm}.}

\item{data}{matrix; Data matrix for all X, see \code{?gmm::gmm}.}
}
\value{
Returns moment the moment conditions
}
\description{
Auxiliary functions for GMM estimation - Poisson
}
