% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{poisson_resid}
\alias{poisson_resid}
\title{Getting residuals for gmm poisson}
\usage{
poisson_resid(gmmo)
}
\arguments{
\item{gmmo}{gmm() output}
}
\value{
residuals of the gmm()
}
\description{
Getting residuals for gmm poisson
}
