% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qapCSS.R
\name{print.QAPCSS}
\alias{print.QAPCSS}
\title{Print QAP() results}
\usage{
\method{print}{QAPCSS}(x, ...)
}
\arguments{
\item{x}{An object of class \code{QAPCSS}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print QAP() results
}
