% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.QAPGLM.R
\name{print.QAPGLM}
\alias{print.QAPGLM}
\title{Print QAP() results for generalized linear regressions}
\usage{
\method{print}{QAPGLM}(x, ..., print_b = FALSE, print_random = FALSE)
}
\arguments{
\item{x}{An object of class \code{QAPGLM}.}

\item{...}{Additional arguments (ignored).}

\item{print_b}{Logical; also show parameter-based p-values?}

\item{print_random}{Logical; show random intercepts?}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print QAP() results for generalized linear regressions
}
