% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{qap_from_igraph}
\alias{qap_from_igraph}
\title{Adjacency matrix from an igraph object}
\usage{
qap_from_igraph(g, attr = NULL)
}
\arguments{
\item{g}{An \pkg{igraph} graph.}

\item{attr}{Character or NULL; edge attribute to use as the weight.}
}
\value{
A numeric matrix with vertex names as dimnames.
}
\description{
Adjacency matrix from an igraph object
}
\keyword{internal}
