% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{qap_from_network}
\alias{qap_from_network}
\title{Adjacency matrix from a network object}
\usage{
qap_from_network(g, attr = NULL)
}
\arguments{
\item{g}{A \pkg{network} object.}

\item{attr}{Character or NULL; edge attribute to use as the weight.}
}
\value{
A numeric matrix.
}
\description{
Adjacency matrix from a network object
}
\keyword{internal}
