% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{qap_is_comparison}
\alias{qap_is_comparison}
\title{Is this a comparison model?}
\usage{
qap_is_comparison(object)
}
\arguments{
\item{object}{A QAP fit.}
}
\value{
Logical.
}
\description{
Is this a comparison model?
}
\keyword{internal}
