% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_fun.R
\name{zip_resid}
\alias{zip_resid}
\title{Residuals from a GMM zero-inflated Poisson fit}
\usage{
zip_resid(gmmo)
}
\arguments{
\item{gmmo}{gmm() output from zip_moments.}
}
\value{
Numeric vector of residuals.
}
\description{
Residuals from a GMM zero-inflated Poisson fit
}
\keyword{internal}
