% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_qap_estimates.R
\name{combine_qap_estimates}
\alias{combine_qap_estimates}
\title{Combine multiple estimates of the same QAP model}
\usage{
combine_qap_estimates(res, res2 = NULL)
}
\arguments{
\item{res}{list; either a \code{list} of \code{class} \code{QAPcss}, \code{QAPregression}, or \code{QAPglm}, or a \code{list} of multiple model outputs that should all be combined.}

\item{res2}{list; if \code{res} is a single model output, \code{res2} is the other model output that should be combined with the first.}
}
\value{
Returns a \code{list} of the \code{class} of the combined model output.
}
\description{
The function combines multiple outputs of \code{QAPglm()} or \code{QAPcss()} estimated on the same data with the same predictors. Models are weighted by \code{reps} when combining, thus, \code{reps} does not need to be the same for all models in \code{res} and \code{res2}.
}
