% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_base.R
\name{fit_base}
\alias{fit_base}
\title{Internal Auxiliary function to fit the base model in QAPcss()}
\usage{
fit_base(mod, rand, family, pred, nx, y, use_robust_errors)
}
\arguments{
\item{mod}{formula; model to be estimated.}

\item{rand}{logical; are there any random intercepts requested?}

\item{family}{character; model family to be estimated.}

\item{pred}{matrix; matrix of predictors.}

\item{nx}{numeric; number of predictors.}

\item{y}{numeric or character; dependent data vector}

\item{use_robust_errors}{logical; should standard errors be corrected by HC3?}
}
\value{
Returns fit for the baseline model
}
\description{
Internal Auxiliary function to fit the base model in QAPcss()
}
