% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_perm.R
\name{fit_perm}
\alias{fit_perm}
\title{Internal Auxiliary function to a permutation model for QAPcss()}
\usage{
fit_perm(
  family. = family.,
  predx = pred,
  ref = reference.,
  xi. = xi.,
  mod. = mod.,
  fitx = fit.,
  rand = rand,
  use_robust_errors. = use_robust_errors
)
}
\arguments{
\item{family.}{character; model family to be estimated.}

\item{predx}{data.frame; predictor matrix.}

\item{ref}{character; reference category if \code{family. == 'multinom'}.}

\item{xi.}{numeric; number of the predictor currently being permuted.}

\item{mod.}{formula; model to be estimated.}

\item{fitx}{list; output from \code{fit_base()}.}

\item{rand}{logical; are there any random intercepts requested?}

\item{use_robust_errors.}{logical; should standard errors be corrected by HC3?}

\item{nx}{numeric; number of predictors.}
}
\value{
Returns if a current permutation is more or less extreme than observed.
}
\description{
Internal Auxiliary function to a permutation model for QAPcss()
}
