% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_tab.R
\name{glm_tab}
\alias{glm_tab}
\title{Internal auxiliary function to help plot QAPcss() output}
\usage{
glm_tab(
  x,
  print_b = print_b,
  print_random = print_random,
  comp,
  nullhyp,
  groups
)
}
\arguments{
\item{x}{list; an \code{R} object of class \code{QAPCSS} returned by \code{QAPcss(...)}.}

\item{print_b}{logical; Shall p-values derived from parameter comparisons also be returned or only those derived from T-values (which is more accurate in most cases). Default FALSE}

\item{print_random}{logical; Shall all random intercepts be returned? Default FALSE}

\item{comp}{list; see \code{QAPcss(...)}}

\item{nullhyp}{character or numeric; see \code{QAPcss(...)}}

\item{groups}{character or numeric; see \code{QAPcss(...)}}
}
\value{
Prints the results table.
}
\description{
Internal auxiliary function to help plot QAPcss() output
}
