% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_tab.R
\name{glm_tab}
\alias{glm_tab}
\title{Internal auxiliary function to help plot QAPcss() output}
\usage{
glm_tab(x, comp)
}
\arguments{
\item{x}{list; an \code{R} object of class \code{QAPCSS} returned by \code{QAPcss(...)}.}

\item{comp}{list; see \code{QAPcss(...)}}
}
\value{
Prints the results table.
}
\description{
Internal auxiliary function to help plot QAPcss() output
}
