% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC3.R
\name{HC3}
\alias{HC3}
\title{Perform HC3 correction for heteroskedasticity}
\usage{
HC3(X, e)
}
\arguments{
\item{X}{matrix; Matrix of predictors}

\item{e}{numeric; vector of residuals}
}
\value{
Returns adjusted covariance matrix.
}
\description{
Perform HC3 correction for heteroskedasticity
}
