% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_css_data.R
\name{make_css_data}
\alias{make_css_data}
\title{Internal auxiliary function to shape data into internal QAPcss() format}
\usage{
make_css_data(y, x, nets, RIO, rio, diag, mode)
}
\arguments{
\item{y}{array or list; see \code{QAPcss()}}

\item{x}{array or list; see \code{QAPcss()}}

\item{nets}{numeric; see \code{QAPcss()}}

\item{RIO}{logical; see \code{QAPcss()}}

\item{rio}{array or list; see \code{QAPcss()}}

\item{diag}{logical; see \code{QAPcss()}}

\item{mode}{character; see \code{QAPcss()}}
}
\value{
Returns vectorized CSS data
}
\description{
Internal auxiliary function to shape data into internal QAPcss() format
}
