% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qap_data.R
\name{make_qap_data}
\alias{make_qap_data}
\title{Internal auxiliary function to make QAPglm() data}
\usage{
make_qap_data(y, x, g, RIO, diag, mode, net, perm = FALSE, xi = NULL)
}
\arguments{
\item{y}{matrix; same as in \code{QAPglm()}}

\item{x}{matrix or list; same as in \code{QAPglm()}}

\item{g}{vector; created in same as in \code{QAPglm()} from groups}

\item{RIO}{matrix or list; same as \code{random_intercept_other} in \code{QAPglm()}}

\item{diag}{logical; same as in \code{QAPglm()}}

\item{mode}{character; same as in \code{QAPglm()}}

\item{net}{integer; internal parameter to create random intercept for each network in \code{y}}

\item{perm}{logical; should a permutation be performed}

\item{xi}{integer; either \code{NULL} or the number of the residualized variable in \code{x}}
}
\value{
data in the format to be used in \code{QAPglm()} or \code{QAPglmPermEst}
}
\description{
Internal auxiliary function to make QAPglm() data
}
