% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qapCSS.R
\docType{methods}
\name{print.QAPCSS}
\alias{print.QAPCSS}
\title{Print QAPcss() results}
\usage{
\method{print}{QAPCSS}(x, print_b = FALSE, print_random = FALSE, ...)
}
\arguments{
\item{x}{list; an \code{R} object of class \code{QAPCSS} returned by \code{QAPcss()}.}

\item{print_b}{logical; Shall p-values derived from parameter comparisons also be returned or only those derived from T-values (which is more accurate in most cases). Default FALSE}

\item{print_random}{logical; Shall all random intercepts be returned? Default FALSE}

\item{...}{Potential other parameters to be passed to lower level functions}
}
\value{
Prints a results table for a \code{QAPcss()} model.
}
\description{
This function prints results for \code{QAPcsss()}.
}
