% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qapCSS.R
\docType{methods}
\name{print.QAPCSS}
\alias{print.QAPCSS}
\title{Print QAPcss() results}
\usage{
\method{print}{QAPCSS}(x, ...)
}
\arguments{
\item{x}{list; an \code{R} object of class \code{QAPCSS} returned by \code{QAPcss()}.}

\item{...}{Potential other parameters to be passed to lower level functions}
}
\value{
Prints a results table for a \code{QAPcss()} model.
}
\description{
This function prints results for \code{QAPcsss()}.
}
