% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.QAPGLM.R
\docType{methods}
\name{print.QAPGLM}
\alias{print.QAPGLM}
\title{Print QAPglm() results for generalized linear regressions}
\usage{
\method{print}{QAPGLM}(x, ..., print_b = FALSE, print_random = FALSE)
}
\arguments{
\item{x}{list; an \code{R} object of class \code{QAPGLM} returned by \code{QAPglm(...), where is not 'gaussian'}.}

\item{...}{Potential other parameters to be passed to lower level functions}

\item{print_b}{logical; Shall p-values derived from parameter comparisons also be returned or only those derived from T-values (which is more accurate in most cases). Default FALSE}

\item{print_random}{logical; Shall all random intercepts be returned? Default FALSE}
}
\value{
Prints a results table for a \code{QAPGLM} model.
}
\description{
This function prints results for a \code{QAPglm()} call where \code{family} is not \code{'gaussian'}, for \code{family = 'gaussian'} use \code{print.QAPRegression()} instead.
}
