% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.QAPRegression.R
\name{print.QAPRegression}
\alias{print.QAPRegression}
\title{Print QAPglm() results for linear regression}
\usage{
\method{print}{QAPRegression}(x, ..., print_b = FALSE, print_random = FALSE)
}
\arguments{
\item{x}{An object of class \code{QAPRegression}.}

\item{...}{Additional arguments (ignored).}

\item{print_b}{Logical; also show parameter-based p-values?}

\item{print_random}{Logical; show random intercepts?}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print QAPglm() results for linear regression
}
