% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.QAPRegression.R
\docType{methods}
\name{print.QAPRegression}
\alias{print.QAPRegression}
\title{Print QAPglm() results for linear regression or linear probability model}
\usage{
\method{print}{QAPRegression}(x, ..., print_b = FALSE, print_random = FALSE)
}
\arguments{
\item{x}{an \code{R} object of class \code{QAPRegression} returned by \code{QAPglm(..., family = 'gaussian')}.}

\item{...}{Potential other parameters to be passed to lower level functions}

\item{print_b}{Shall p-values derived from parameter comparisons also be returned or only those derived from T-values (which is more accurate in most cases). Default FALSE}

\item{print_random}{Shall all random intercepts be returned? By default FALSE}
}
\value{
Prints a results table for a \code{QAPRegression} model.
}
\description{
This function prints results for a \code{QAPglm(..., family = 'gaussian')} call.
}
