% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAPcssPermEst.R
\name{QAPcssPermEst}
\alias{QAPcssPermEst}
\title{Internal auxiliary function to estimate one CSS permutation}
\usage{
QAPcssPermEst(
  i,
  y.,
  x.,
  g. = NULL,
  mode.,
  diag.,
  rand. = rand,
  family.,
  groups.,
  fit.,
  comp.,
  RIO.,
  use_robust_errors.,
  xi. = NULL,
  xRm. = NULL,
  reference. = NULL,
  mod.
)
}
\arguments{
\item{i}{NULL; iterator of \code{parLapply()}}

\item{y.}{array or list; same as \code{y} in \code{QAPcss()}}

\item{x.}{array or list; same as \code{x} in \code{QAPcss()}}

\item{g.}{vector; similar to as \code{groups} in \code{QAPcss()}}

\item{mode.}{character; same as \code{mode} in \code{QAPcss()}}

\item{diag.}{logical; same as \code{diag} in \code{QAPcss()}}

\item{rand.}{logical; TRUE if any random intercepts are requested}

\item{family.}{character; same as \code{family} in \code{QAPcss()}}

\item{groups.}{vector; same as \code{groups} in \code{QAPcss()}}

\item{fit.}{glm, lm, lmer, glmer, or multinom; Estimate on unpermuted data from \code{fit_base}}

\item{comp.}{list; same as \code{comparison} in \code{QAPcss()}}

\item{RIO.}{array or list; same as \code{random_intercept_other} in \code{QAPcss()}}

\item{use_robust_errors.}{logical; same as \code{use_robust_errors} in \code{QAPcss()}}

\item{xi.}{integer; either \code{NULL} or the number of the residualized variable in \code{x}}

\item{xRm.}{array; residualized \code{array} for predictor \code{xi}}

\item{reference.}{numeric or character; same as \code{reference} in \code{QAPcss()}}

\item{mod.}{formula; model to be estimated.}
}
\value{
Returns permutation results for one CSS permutation.
}
\description{
Internal auxiliary function to estimate one CSS permutation
}
