% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAPglmPermEst.R
\name{QAPglmPermEst}
\alias{QAPglmPermEst}
\title{Internal auxiliary function to estimate one permutation}
\usage{
QAPglmPermEst(
  i,
  y.,
  mode.,
  diag.,
  mod.,
  family.,
  groups.,
  fit.,
  RIO.,
  use_robust_errors.,
  xi. = NULL,
  xRm. = NULL,
  same_x_4_all_y.,
  rand.
)
}
\arguments{
\item{i}{NULL; iterator of parLapply}

\item{y.}{matrix or list; same as \code{y} in \code{QAPglm()}}

\item{mode.}{character; same as \code{mode} in \code{QAPglm()}}

\item{diag.}{logical; same as \code{diag} in \code{QAPglm()}}

\item{mod.}{formula; model to be estimated.}

\item{family.}{character; same as \code{family} in \code{QAPglm()}}

\item{groups.}{vector; same as \code{groups} in \code{QAPglm()}}

\item{fit.}{glm, lm, lmer, or glmer; Estimate on unpermuted data.}

\item{RIO.}{matrix or list; same as \code{random_intercept_other} in \code{QAPglm()}}

\item{use_robust_errors.}{logical; same as \code{use_robust_errors} in \code{QAPglm()}}

\item{xi.}{integer; either \code{NULL} or the number of the residualized variable in \code{x}}

\item{xRm.}{matrix or list; same as \code{x} in \code{QAPglm()}}

\item{same_x_4_all_y.}{matrix or list; same as \code{y} in \code{QAPglm()}}

\item{rand.}{logical; TRUE if any random intercepts are requested}
}
\value{
Results for one permutation.
}
\description{
Internal auxiliary function to estimate one permutation
}
