% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals_to_array}
\alias{residuals_to_array}
\title{Put residuals back into 3D array form (CSS)}
\usage{
residuals_to_array(
  xR,
  original_array,
  valid,
  pred,
  large = FALSE,
  valid_list = NULL
)
}
\arguments{
\item{xR}{Numeric vector of residuals.}

\item{original_array}{The original 3D array (or list of arrays).}

\item{valid}{Logical array of valid positions.}

\item{pred}{Data frame with \code{nv}.}

\item{large}{Logical; TRUE if y is a list of arrays.}

\item{valid_list}{List of valid arrays (when large = TRUE).}
}
\value{
An array (or list of arrays) with residuals.
}
\description{
Put residuals back into 3D array form (CSS)
}
\keyword{internal}
