% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPerm.R
\name{RMPerm}
\alias{RMPerm}
\title{Internal auxiliary function to permute matrices or arrays}
\usage{
RMPerm(m, groups = NULL, CSS = FALSE)
}
\arguments{
\item{m}{matrix, array, or list; a \code{matrix}, \code{array}, or \code{list} thereof to be permuted}

\item{groups}{vector; \code{vector} with respective grouping (see \code{?QAPglm()} and \code{?QAPcss()})}

\item{CSS}{logical; is it a CSS structure}
}
\value{
a permuted \code{matrix} or \code{array} or \code{list} thereof as provided in \code{m}
}
\description{
Internal auxiliary function to permute matrices or arrays
}
