% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_permutations}
\alias{run_permutations}
\title{Run permutations in parallel using future}
\usage{
run_permutations(reps, FUN, ..., p = NULL)
}
\arguments{
\item{reps}{Number of permutations.}

\item{FUN}{Function to run per permutation.}

\item{...}{Additional arguments passed to FUN.}

\item{p}{A \code{progressr::progressor} object or NULL.}
}
\value{
List of permutation results.
}
\description{
Run permutations in parallel using future
}
\keyword{internal}
